requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
brotli==1.2.0
brotlipy==0.7.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
coverage==7.13.1
distro==1.9.0
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.2
fsspec==2025.12.0
google-auth==2.47.0
google-genai==1.56.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
llvmlite==0.46.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
mutagen==1.47.0
networkx==3.6.1
numba==0.63.1
numby==0.2.0
numpy==2.3.5
openai-whisper==20250625
packaging==25.0
pluggy==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dotenv==1.2.1
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
torch==2.9.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
websockets==15.0.1
yt-dlp==2025.12.8
yt-dlp-ejs==0.3.2